Separate logic into hepler folder
