+++ src/Scenes/GameScene.js
@@ -1,113 +1,7 @@
 import LocalStorage from '../Objects/localStorage';
-
-class Laser extends Phaser.Physics.Arcade.Sprite {
-  constructor(scene, x, y) {
-    super(scene, x, y, 'playerAttack', 30);
-  }
-
-  fire(x, y) {
-    this.body.reset(x, y);
-    this.setActive(true);
-    this.setVisible(true);
-    this.setVelocityX(700);
-    this.body.enable = true
-  }
-}
-
-class Enemy extends Phaser.Physics.Arcade.Sprite {
-  constructor(scene, x, y) {
-    super(scene, x, y, 'enemy', 10);
-  }
-
-  enemyPosition(x, y) {
-    this.body.reset(x, y);
-    this.setActive(true);
-    this.setVisible(true);
-    this.flipX = true
-    this.setScale(1.4, 1.4)
-    this.anims.play('enemy')
-    this.body.enable = true
-  }
-}
-
-class LaserGroup extends Phaser.Physics.Arcade.Group {
-  constructor(scene) {
-    super(scene.physics.world, scene);
-
-    this.createMultiple({
-      frameQuantity: 1,
-      key: 'playerAttack',
-      active: false,
-      visible: false,
-      classType: Laser
-    });
-  }
-
-  fireBullet(x, y) {
-    const laser = this.getFirstDead(false);
-
-    if (laser) {
-      laser.fire(x, y);
-    }
-  }
-}
-
-class EnemyGroup extends Phaser.Physics.Arcade.Group {
-  constructor(scene) {
-    super(scene.physics.world, scene);
-
-    this.createMultiple({
-      frameQuantity: 30,
-      key: 'enemy',
-      active: false,
-      visible: false,
-      classType: Enemy
-    });
-  }
-
-  createEnemy(x, y) {
-    const enemy = this.getFirstDead(false);
-
-    if (enemy) {
-      enemy.enemyPosition(x, y);
-    }
-  }
-}
-
-class EnemyAttack extends Phaser.Physics.Arcade.Sprite {
-  constructor(scene, x, y) {
-    super(scene, x, y, 'enemyAttack', 10);
-  }
-
-  enemyPosition(x, y) {
-    this.body.reset(x, y);
-    this.setActive(true);
-    this.setVisible(true);
-    this.body.enable = true
-  }
-}
-
-class EnemyAttackGroup extends Phaser.Physics.Arcade.Group {
-  constructor(scene) {
-    super(scene.physics.world, scene);
-
-    this.createMultiple({
-      frameQuantity: 100,
-      key: 'enemyAttack',
-      active: false,
-      visible: false,
-      loop: true,
-      classType: EnemyAttack
-    });
-  }
-  
-  createEnemyAttack(x, y, player, scenes, playerAttackGroup ) {
-    const enemy = this.getFirstDead(false);
-    if (enemy) {
-      enemy.enemyPosition(x, y, player, scenes, playerAttackGroup )
-    };
-  }
-}
+import EnemyGroup from '../helper/enemyGroup';
+import EnemyAttackGroup from '../helper/enemyAttackGroup';
+import LaserGroup from '../helper/playerAttackGroup';
 
 var score = 0;
 let checkFunction = true
